% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/genomicsFunctions.R
\name{DT2GR}
\alias{DT2GR}
\title{Convert a data table to a GRanges object
By default it checks that the GRanges object
has the chromosomes in si}
\usage{
DT2GR(dt, seqinf = si, checksi = TRUE)
}
\description{
Convert a data table to a GRanges object
By default it checks that the GRanges object
has the chromosomes in si
}

