% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/genomicsFunctions.R
\name{splitByCol}
\alias{splitByCol}
\title{This function takes in a table, a bare column name, and a string indicating
the seperator. It then puts each seperated value on it's own row, with
All other columns duplicated}
\usage{
splitByCol(x, colname, split)
}
\arguments{
\item{x}{- a table or Granges object}

\item{colname}{- a column name, bare}

\item{split}{- a character to split the column by}
}
\value{
a table where the concatenated column has been split into individual rows
splitByCol(tbl,columntosplit,seperatorstring)
}
\description{
This function takes in a table, a bare column name, and a string indicating
the seperator. It then puts each seperated value on it's own row, with
All other columns duplicated
}

